Split LOI write trigger into create, update and delete handlers
