+++ functions/src/on-update-loi.ts
@@ -1,0 +1,35 @@
+/**
+ * Copyright 2026 The Ground Authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the 'License');
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an 'AS IS' BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import {
+  Change,
+  DocumentSnapshot,
+  FirestoreEvent,
+} from 'firebase-functions/v2/firestore';
+import { broadcastUpdate } from './common/broadcast';
+
+export async function onUpdateLoiHandler(
+  event: FirestoreEvent<Change<DocumentSnapshot> | undefined>
+) {
+  const { surveyId, loiId } = event.params;
+
+  if (!surveyId || !loiId) return;
+
+  return broadcastUpdate(
+    { type: 'loi', surveyId, loiId, deleted: false },
+    event.time
+  );
+}
